+++ wow_login_messages/src/helper.rs
@@ -40,7 +40,7 @@ pub fn expect_client_message<M: ClientMessage, R: std::io::Read>(
 
     // Unable to match on associated const M::OPCODE, so we do if
     if opcode == M::OPCODE {
-        let m = M::read(&mut r);
+        let m = M::read::<&mut R, crate::private::Internal>(&mut r);
         match m {
             Ok(m) => Ok(m),
             Err(e) => Err(e.into()),
@@ -62,7 +62,7 @@ pub async fn tokio_expect_client_message<
 
     // Unable to match on associated const M::OPCODE, so we do if
     if opcode == M::OPCODE {
-        let m = M::tokio_read(&mut r).await;
+        let m = M::tokio_read::<&mut R, crate::private::Internal>(&mut r).await;
         match m {
             Ok(m) => Ok(m),
             Err(e) => Err(e.into()),
@@ -84,7 +84,7 @@ pub async fn astd_expect_client_message<
 
     // Unable to match on associated const M::OPCODE, so we do if
     if opcode == M::OPCODE {
-        let m = M::astd_read(&mut r).await;
+        let m = M::astd_read::<&mut R, crate::private::Internal>(&mut r).await;
         match m {
             Ok(m) => Ok(m),
             Err(e) => Err(e.into()),
@@ -116,7 +116,7 @@ pub fn expect_server_message<M: ServerMessage, R: std::io::Read>(
 
     // Unable to match on associated const M::OPCODE, so we do if
     if opcode == M::OPCODE {
-        let m = M::read(&mut r);
+        let m = M::read::<&mut R, crate::private::Internal>(&mut r);
         match m {
             Ok(m) => Ok(m),
             Err(e) => Err(e.into()),
@@ -138,7 +138,7 @@ pub async fn tokio_expect_server_message<
 
     // Unable to match on associated const M::OPCODE, so we do if
     if opcode == M::OPCODE {
-        let m = M::tokio_read(&mut r).await;
+        let m = M::tokio_read::<&mut R, crate::private::Internal>(&mut r).await;
         match m {
             Ok(m) => Ok(m),
             Err(e) => Err(e.into()),
@@ -160,7 +160,7 @@ pub async fn astd_expect_server_message<
 
     // Unable to match on associated const M::OPCODE, so we do if
     if opcode == M::OPCODE {
-        let m = M::astd_read(&mut r).await;
+        let m = M::astd_read::<&mut R, crate::private::Internal>(&mut r).await;
         match m {
             Ok(m) => Ok(m),
             Err(e) => Err(e.into()),
@@ -219,10 +219,10 @@ pub fn read_initial_message<R: std::io::Read>(
     let opcode = read_u8_le(&mut r)?;
     match opcode {
         CMD_AUTH_LOGON_CHALLENGE_Client::OPCODE => Ok(InitialMessage::Logon(
-            CMD_AUTH_LOGON_CHALLENGE_Client::read(&mut r)?,
+            CMD_AUTH_LOGON_CHALLENGE_Client::read::<&mut R, crate::private::Internal>(&mut r)?,
         )),
         CMD_AUTH_RECONNECT_CHALLENGE_Client::OPCODE => Ok(InitialMessage::Reconnect(
-            CMD_AUTH_RECONNECT_CHALLENGE_Client::read(&mut r)?,
+            CMD_AUTH_RECONNECT_CHALLENGE_Client::read::<&mut R, crate::private::Internal>(&mut r)?,
         )),
         opcode => Err(ExpectedOpcodeError::Opcode(opcode as u32)),
     }
@@ -236,10 +236,10 @@ pub async fn tokio_read_initial_message<R: tokio::io::AsyncReadExt + Unpin + Sen
     let opcode = crate::util::tokio_read_u8_le(&mut r).await?;
     match opcode {
         CMD_AUTH_LOGON_CHALLENGE_Client::OPCODE => Ok(InitialMessage::Logon(
-            CMD_AUTH_LOGON_CHALLENGE_Client::tokio_read(&mut r).await?,
+            CMD_AUTH_LOGON_CHALLENGE_Client::tokio_read::<&mut R, crate::private::Internal>(&mut r).await?,
         )),
         CMD_AUTH_RECONNECT_CHALLENGE_Client::OPCODE => Ok(InitialMessage::Reconnect(
-            CMD_AUTH_RECONNECT_CHALLENGE_Client::tokio_read(&mut r).await?,
+            CMD_AUTH_RECONNECT_CHALLENGE_Client::tokio_read::<&mut R, crate::private::Internal>(&mut r).await?,
         )),
         opcode => Err(ExpectedOpcodeError::Opcode(opcode as u32)),
     }
@@ -253,10 +253,10 @@ pub async fn astd_read_initial_message<R: async_std::io::ReadExt + Unpin + Send>
     let opcode = crate::util::astd_read_u8_le(&mut r).await?;
     match opcode {
         CMD_AUTH_LOGON_CHALLENGE_Client::OPCODE => Ok(InitialMessage::Logon(
-            CMD_AUTH_LOGON_CHALLENGE_Client::astd_read(&mut r).await?,
+            CMD_AUTH_LOGON_CHALLENGE_Client::astd_read::<&mut R, crate::private::Internal>(&mut r).await?,
         )),
         CMD_AUTH_RECONNECT_CHALLENGE_Client::OPCODE => Ok(InitialMessage::Reconnect(
-            CMD_AUTH_RECONNECT_CHALLENGE_Client::astd_read(&mut r).await?,
+            CMD_AUTH_RECONNECT_CHALLENGE_Client::astd_read::<&mut R, crate::private::Internal>(&mut r).await?,
         )),
         opcode => Err(ExpectedOpcodeError::Opcode(opcode as u32)),
     }