+++ wow_login_messages/src/lib.rs
@@ -133,19 +133,19 @@ pub const DEFAULT_PORT: u16 = 3724;
 ///
 /// This trait also has a bunch of hidden functions that are necessary for the [`helper`](crate::helper)
 /// and [`opcodes`](crate::version_2::opcodes) modules to work.
-pub trait ServerMessage: Sized {
+pub trait ServerMessage: Sized + private::Sealed {
     #[doc(hidden)]
     const OPCODE: u8;
 
     #[doc(hidden)]
-    fn read<R: std::io::Read>(r: R) -> Result<Self, crate::errors::ParseError>;
+    fn read<R: std::io::Read, I: private::Sealed>(r: R) -> Result<Self, crate::errors::ParseError>;
 
     #[cfg(feature = "sync")]
     fn write<W: std::io::Write>(&self, w: W) -> Result<(), std::io::Error>;
 
     #[doc(hidden)]
     #[cfg(feature = "async-std")]
-    fn astd_read<'async_trait, R>(
+    fn astd_read<'async_trait, R, I: private::Sealed>(
         r: R,
     ) -> Pin<Box<dyn Future<Output = Result<Self, crate::errors::ParseError>> + Send + 'async_trait>>
     where
@@ -164,7 +164,7 @@ pub trait ServerMessage: Sized {
 
     #[doc(hidden)]
     #[cfg(feature = "tokio")]
-    fn tokio_read<'async_trait, R>(
+    fn tokio_read<'async_trait, R, I: private::Sealed>(
         r: R,
     ) -> Pin<Box<dyn Future<Output = Result<Self, crate::errors::ParseError>> + Send + 'async_trait>>
     where
@@ -192,19 +192,19 @@ pub trait ServerMessage: Sized {
 ///
 /// This trait also has a bunch of hidden functions that are necessary for the [`helper`](crate::helper)
 /// and [`opcodes`](crate::version_2::opcodes) modules to work.
-pub trait ClientMessage: Sized {
+pub trait ClientMessage: Sized + private::Sealed {
     #[doc(hidden)]
     const OPCODE: u8;
 
     #[doc(hidden)]
-    fn read<R: std::io::Read>(r: R) -> Result<Self, crate::errors::ParseError>;
+    fn read<R: std::io::Read, I: private::Sealed>(r: R) -> Result<Self, crate::errors::ParseError>;
 
     #[cfg(feature = "sync")]
     fn write<W: std::io::Write>(&self, w: W) -> Result<(), std::io::Error>;
 
     #[doc(hidden)]
     #[cfg(feature = "async-std")]
-    fn astd_read<'async_trait, R>(
+    fn astd_read<'async_trait, R, I: private::Sealed>(
         r: R,
     ) -> Pin<Box<dyn Future<Output = Result<Self, crate::errors::ParseError>> + Send + 'async_trait>>
     where
@@ -223,7 +223,7 @@ pub trait ClientMessage: Sized {
 
     #[doc(hidden)]
     #[cfg(feature = "tokio")]
-    fn tokio_read<'async_trait, R>(
+    fn tokio_read<'async_trait, R, I: private::Sealed>(
         r: R,
     ) -> Pin<Box<dyn Future<Output = Result<Self, crate::errors::ParseError>> + Send + 'async_trait>>
     where
@@ -239,4 +239,13 @@ pub trait ClientMessage: Sized {
         W: 'async_trait + AsyncWriteExt + Unpin + Send,
         'life0: 'async_trait,
         Self: 'async_trait;
+}
+
+#[doc(hidden)]
+pub(crate) mod private {
+    pub trait Sealed {}
+
+    pub(crate) struct Internal;
+
+    impl Sealed for Internal {}
 }