+++ wow_login_messages/src/logon/all/cmd_auth_logon_challenge_client.rs
@@ -99,10 +99,12 @@ impl CMD_AUTH_LOGON_CHALLENGE_Client {
     }
 }
 
+impl crate::private::Sealed for CMD_AUTH_LOGON_CHALLENGE_Client {}
+
 impl ClientMessage for CMD_AUTH_LOGON_CHALLENGE_Client {
     const OPCODE: u8 = 0x00;
 
-    fn read<R: std::io::Read>(mut r: R) -> std::result::Result<Self, crate::errors::ParseError> {
+    fn read<R: std::io::Read, I: crate::private::Sealed>(mut r: R) -> std::result::Result<Self, crate::errors::ParseError> {
         // protocol_version: u8
         let protocol_version = crate::util::read_u8_le(&mut r)?;
 
@@ -159,7 +161,7 @@ impl ClientMessage for CMD_AUTH_LOGON_CHALLENGE_Client {
     }
 
     #[cfg(feature = "tokio")]
-    fn tokio_read<'async_trait, R>(
+    fn tokio_read<'async_trait, R, I: crate::private::Sealed>(
         mut r: R,
     ) -> core::pin::Pin<Box<
         dyn core::future::Future<Output = std::result::Result<Self, crate::errors::ParseError>>
@@ -238,7 +240,7 @@ impl ClientMessage for CMD_AUTH_LOGON_CHALLENGE_Client {
     }
 
     #[cfg(feature = "async-std")]
-    fn astd_read<'async_trait, R>(
+    fn astd_read<'async_trait, R, I: crate::private::Sealed>(
         mut r: R,
     ) -> core::pin::Pin<Box<
         dyn core::future::Future<Output = std::result::Result<Self, crate::errors::ParseError>>