+++ wow_login_messages/src/logon/version_3/cmd_auth_logon_challenge_server.rs
@@ -125,10 +125,12 @@ impl CMD_AUTH_LOGON_CHALLENGE_Server {
     }
 }
 
+impl crate::private::Sealed for CMD_AUTH_LOGON_CHALLENGE_Server {}
+
 impl ServerMessage for CMD_AUTH_LOGON_CHALLENGE_Server {
     const OPCODE: u8 = 0x00;
 
-    fn read<R: std::io::Read>(mut r: R) -> std::result::Result<Self, crate::errors::ParseError> {
+    fn read<R: std::io::Read, I: crate::private::Sealed>(mut r: R) -> std::result::Result<Self, crate::errors::ParseError> {
         // protocol_version: u8
         let _protocol_version = crate::util::read_u8_le(&mut r)?;
         // protocol_version is expected to always be 0 (0)
@@ -245,7 +247,7 @@ impl ServerMessage for CMD_AUTH_LOGON_CHALLENGE_Server {
     }
 
     #[cfg(feature = "tokio")]
-    fn tokio_read<'async_trait, R>(
+    fn tokio_read<'async_trait, R, I: crate::private::Sealed>(
         mut r: R,
     ) -> core::pin::Pin<Box<
         dyn core::future::Future<Output = std::result::Result<Self, crate::errors::ParseError>>
@@ -384,7 +386,7 @@ impl ServerMessage for CMD_AUTH_LOGON_CHALLENGE_Server {
     }
 
     #[cfg(feature = "async-std")]
-    fn astd_read<'async_trait, R>(
+    fn astd_read<'async_trait, R, I: crate::private::Sealed>(
         mut r: R,
     ) -> core::pin::Pin<Box<
         dyn core::future::Future<Output = std::result::Result<Self, crate::errors::ParseError>>