+++ wow_login_messages/src/logon/version_3/cmd_xfer_cancel.rs
@@ -19,10 +19,12 @@ impl CMD_XFER_CANCEL {
     }
 }
 
+impl crate::private::Sealed for CMD_XFER_CANCEL {}
+
 impl ClientMessage for CMD_XFER_CANCEL {
     const OPCODE: u8 = 0x34;
 
-    fn read<R: std::io::Read>(mut r: R) -> std::result::Result<Self, crate::errors::ParseError> {
+    fn read<R: std::io::Read, I: crate::private::Sealed>(mut r: R) -> std::result::Result<Self, crate::errors::ParseError> {
         Ok(Self {
         })
     }
@@ -35,7 +37,7 @@ impl ClientMessage for CMD_XFER_CANCEL {
     }
 
     #[cfg(feature = "tokio")]
-    fn tokio_read<'async_trait, R>(
+    fn tokio_read<'async_trait, R, I: crate::private::Sealed>(
         mut r: R,
     ) -> core::pin::Pin<Box<
         dyn core::future::Future<Output = std::result::Result<Self, crate::errors::ParseError>>
@@ -70,7 +72,7 @@ impl ClientMessage for CMD_XFER_CANCEL {
     }
 
     #[cfg(feature = "async-std")]
-    fn astd_read<'async_trait, R>(
+    fn astd_read<'async_trait, R, I: crate::private::Sealed>(
         mut r: R,
     ) -> core::pin::Pin<Box<
         dyn core::future::Future<Output = std::result::Result<Self, crate::errors::ParseError>>