+++ wow_login_messages/src/logon/version_3/cmd_xfer_data.rs
@@ -30,10 +30,12 @@ impl CMD_XFER_DATA {
     }
 }
 
+impl crate::private::Sealed for CMD_XFER_DATA {}
+
 impl ServerMessage for CMD_XFER_DATA {
     const OPCODE: u8 = 0x31;
 
-    fn read<R: std::io::Read>(mut r: R) -> std::result::Result<Self, crate::errors::ParseError> {
+    fn read<R: std::io::Read, I: crate::private::Sealed>(mut r: R) -> std::result::Result<Self, crate::errors::ParseError> {
         // size: u16
         let size = crate::util::read_u16_le(&mut r)?;
 
@@ -59,7 +61,7 @@ impl ServerMessage for CMD_XFER_DATA {
     }
 
     #[cfg(feature = "tokio")]
-    fn tokio_read<'async_trait, R>(
+    fn tokio_read<'async_trait, R, I: crate::private::Sealed>(
         mut r: R,
     ) -> core::pin::Pin<Box<
         dyn core::future::Future<Output = std::result::Result<Self, crate::errors::ParseError>>
@@ -107,7 +109,7 @@ impl ServerMessage for CMD_XFER_DATA {
     }
 
     #[cfg(feature = "async-std")]
-    fn astd_read<'async_trait, R>(
+    fn astd_read<'async_trait, R, I: crate::private::Sealed>(
         mut r: R,
     ) -> core::pin::Pin<Box<
         dyn core::future::Future<Output = std::result::Result<Self, crate::errors::ParseError>>