+++ wow_login_messages/src/logon/version_3/opcodes.rs
@@ -33,11 +33,11 @@ impl ServerOpcodeMessage {
     pub fn read<R: std::io::Read>(mut r: R) -> std::result::Result<Self, crate::errors::ExpectedOpcodeError> {
         let opcode = crate::util::read_u8_le(&mut r)?;
         match opcode {
-            0x00 => Ok(Self::CMD_AUTH_LOGON_CHALLENGE(CMD_AUTH_LOGON_CHALLENGE_Server::read(r)?)),
-            0x01 => Ok(Self::CMD_AUTH_LOGON_PROOF(CMD_AUTH_LOGON_PROOF_Server::read(r)?)),
-            0x10 => Ok(Self::CMD_REALM_LIST(CMD_REALM_LIST_Server::read(r)?)),
+            0x00 => Ok(Self::CMD_AUTH_LOGON_CHALLENGE(CMD_AUTH_LOGON_CHALLENGE_Server::read::<R, crate::private::Internal>(r)?)),
+            0x01 => Ok(Self::CMD_AUTH_LOGON_PROOF(CMD_AUTH_LOGON_PROOF_Server::read::<R, crate::private::Internal>(r)?)),
+            0x10 => Ok(Self::CMD_REALM_LIST(CMD_REALM_LIST_Server::read::<R, crate::private::Internal>(r)?)),
             0x30 => Ok(Self::CMD_XFER_INITIATE),
-            0x31 => Ok(Self::CMD_XFER_DATA(CMD_XFER_DATA::read(r)?)),
+            0x31 => Ok(Self::CMD_XFER_DATA(CMD_XFER_DATA::read::<R, crate::private::Internal>(r)?)),
             opcode => Err(crate::errors::ExpectedOpcodeError::Opcode(opcode as u32)),
         }
     }
@@ -46,11 +46,11 @@ impl ServerOpcodeMessage {
     pub async fn tokio_read<R: tokio::io::AsyncReadExt + Unpin + Send>(mut r: R) -> std::result::Result<Self, crate::errors::ExpectedOpcodeError> {
         let opcode = crate::util::tokio_read_u8_le(&mut r).await?;
         match opcode {
-            0x00 => Ok(Self::CMD_AUTH_LOGON_CHALLENGE(CMD_AUTH_LOGON_CHALLENGE_Server::tokio_read(r).await?)),
-            0x01 => Ok(Self::CMD_AUTH_LOGON_PROOF(CMD_AUTH_LOGON_PROOF_Server::tokio_read(r).await?)),
-            0x10 => Ok(Self::CMD_REALM_LIST(CMD_REALM_LIST_Server::tokio_read(r).await?)),
+            0x00 => Ok(Self::CMD_AUTH_LOGON_CHALLENGE(CMD_AUTH_LOGON_CHALLENGE_Server::tokio_read::<R, crate::private::Internal>(r).await?)),
+            0x01 => Ok(Self::CMD_AUTH_LOGON_PROOF(CMD_AUTH_LOGON_PROOF_Server::tokio_read::<R, crate::private::Internal>(r).await?)),
+            0x10 => Ok(Self::CMD_REALM_LIST(CMD_REALM_LIST_Server::tokio_read::<R, crate::private::Internal>(r).await?)),
             0x30 => Ok(Self::CMD_XFER_INITIATE),
-            0x31 => Ok(Self::CMD_XFER_DATA(CMD_XFER_DATA::tokio_read(r).await?)),
+            0x31 => Ok(Self::CMD_XFER_DATA(CMD_XFER_DATA::tokio_read::<R, crate::private::Internal>(r).await?)),
             opcode => Err(crate::errors::ExpectedOpcodeError::Opcode(opcode as u32)),
         }
     }
@@ -59,11 +59,11 @@ impl ServerOpcodeMessage {
     pub async fn astd_read<R: async_std::io::ReadExt + Unpin + Send>(mut r: R) -> std::result::Result<Self, crate::errors::ExpectedOpcodeError> {
         let opcode = crate::util::astd_read_u8_le(&mut r).await?;
         match opcode {
-            0x00 => Ok(Self::CMD_AUTH_LOGON_CHALLENGE(CMD_AUTH_LOGON_CHALLENGE_Server::astd_read(r).await?)),
-            0x01 => Ok(Self::CMD_AUTH_LOGON_PROOF(CMD_AUTH_LOGON_PROOF_Server::astd_read(r).await?)),
-            0x10 => Ok(Self::CMD_REALM_LIST(CMD_REALM_LIST_Server::astd_read(r).await?)),
+            0x00 => Ok(Self::CMD_AUTH_LOGON_CHALLENGE(CMD_AUTH_LOGON_CHALLENGE_Server::astd_read::<R, crate::private::Internal>(r).await?)),
+            0x01 => Ok(Self::CMD_AUTH_LOGON_PROOF(CMD_AUTH_LOGON_PROOF_Server::astd_read::<R, crate::private::Internal>(r).await?)),
+            0x10 => Ok(Self::CMD_REALM_LIST(CMD_REALM_LIST_Server::astd_read::<R, crate::private::Internal>(r).await?)),
             0x30 => Ok(Self::CMD_XFER_INITIATE),
-            0x31 => Ok(Self::CMD_XFER_DATA(CMD_XFER_DATA::astd_read(r).await?)),
+            0x31 => Ok(Self::CMD_XFER_DATA(CMD_XFER_DATA::astd_read::<R, crate::private::Internal>(r).await?)),
             opcode => Err(crate::errors::ExpectedOpcodeError::Opcode(opcode as u32)),
         }
     }
@@ -146,13 +146,13 @@ impl ClientOpcodeMessage {
     pub fn read<R: std::io::Read>(mut r: R) -> std::result::Result<Self, crate::errors::ExpectedOpcodeError> {
         let opcode = crate::util::read_u8_le(&mut r)?;
         match opcode {
-            0x00 => Ok(Self::CMD_AUTH_LOGON_CHALLENGE(CMD_AUTH_LOGON_CHALLENGE_Client::read(r)?)),
-            0x01 => Ok(Self::CMD_AUTH_LOGON_PROOF(CMD_AUTH_LOGON_PROOF_Client::read(r)?)),
-            0x02 => Ok(Self::CMD_AUTH_RECONNECT_CHALLENGE(CMD_AUTH_RECONNECT_CHALLENGE_Client::read(r)?)),
-            0x04 => Ok(Self::CMD_SURVEY_RESULT(CMD_SURVEY_RESULT::read(r)?)),
-            0x10 => Ok(Self::CMD_REALM_LIST(CMD_REALM_LIST_Client::read(r)?)),
+            0x00 => Ok(Self::CMD_AUTH_LOGON_CHALLENGE(CMD_AUTH_LOGON_CHALLENGE_Client::read::<R, crate::private::Internal>(r)?)),
+            0x01 => Ok(Self::CMD_AUTH_LOGON_PROOF(CMD_AUTH_LOGON_PROOF_Client::read::<R, crate::private::Internal>(r)?)),
+            0x02 => Ok(Self::CMD_AUTH_RECONNECT_CHALLENGE(CMD_AUTH_RECONNECT_CHALLENGE_Client::read::<R, crate::private::Internal>(r)?)),
+            0x04 => Ok(Self::CMD_SURVEY_RESULT(CMD_SURVEY_RESULT::read::<R, crate::private::Internal>(r)?)),
+            0x10 => Ok(Self::CMD_REALM_LIST(CMD_REALM_LIST_Client::read::<R, crate::private::Internal>(r)?)),
             0x32 => Ok(Self::CMD_XFER_ACCEPT),
-            0x33 => Ok(Self::CMD_XFER_RESUME(CMD_XFER_RESUME::read(r)?)),
+            0x33 => Ok(Self::CMD_XFER_RESUME(CMD_XFER_RESUME::read::<R, crate::private::Internal>(r)?)),
             0x34 => Ok(Self::CMD_XFER_CANCEL),
             opcode => Err(crate::errors::ExpectedOpcodeError::Opcode(opcode as u32)),
         }
@@ -162,13 +162,13 @@ impl ClientOpcodeMessage {
     pub async fn tokio_read<R: tokio::io::AsyncReadExt + Unpin + Send>(mut r: R) -> std::result::Result<Self, crate::errors::ExpectedOpcodeError> {
         let opcode = crate::util::tokio_read_u8_le(&mut r).await?;
         match opcode {
-            0x00 => Ok(Self::CMD_AUTH_LOGON_CHALLENGE(CMD_AUTH_LOGON_CHALLENGE_Client::tokio_read(r).await?)),
-            0x01 => Ok(Self::CMD_AUTH_LOGON_PROOF(CMD_AUTH_LOGON_PROOF_Client::tokio_read(r).await?)),
-            0x02 => Ok(Self::CMD_AUTH_RECONNECT_CHALLENGE(CMD_AUTH_RECONNECT_CHALLENGE_Client::tokio_read(r).await?)),
-            0x04 => Ok(Self::CMD_SURVEY_RESULT(CMD_SURVEY_RESULT::tokio_read(r).await?)),
-            0x10 => Ok(Self::CMD_REALM_LIST(CMD_REALM_LIST_Client::tokio_read(r).await?)),
+            0x00 => Ok(Self::CMD_AUTH_LOGON_CHALLENGE(CMD_AUTH_LOGON_CHALLENGE_Client::tokio_read::<R, crate::private::Internal>(r).await?)),
+            0x01 => Ok(Self::CMD_AUTH_LOGON_PROOF(CMD_AUTH_LOGON_PROOF_Client::tokio_read::<R, crate::private::Internal>(r).await?)),
+            0x02 => Ok(Self::CMD_AUTH_RECONNECT_CHALLENGE(CMD_AUTH_RECONNECT_CHALLENGE_Client::tokio_read::<R, crate::private::Internal>(r).await?)),
+            0x04 => Ok(Self::CMD_SURVEY_RESULT(CMD_SURVEY_RESULT::tokio_read::<R, crate::private::Internal>(r).await?)),
+            0x10 => Ok(Self::CMD_REALM_LIST(CMD_REALM_LIST_Client::tokio_read::<R, crate::private::Internal>(r).await?)),
             0x32 => Ok(Self::CMD_XFER_ACCEPT),
-            0x33 => Ok(Self::CMD_XFER_RESUME(CMD_XFER_RESUME::tokio_read(r).await?)),
+            0x33 => Ok(Self::CMD_XFER_RESUME(CMD_XFER_RESUME::tokio_read::<R, crate::private::Internal>(r).await?)),
             0x34 => Ok(Self::CMD_XFER_CANCEL),
             opcode => Err(crate::errors::ExpectedOpcodeError::Opcode(opcode as u32)),
         }
@@ -178,13 +178,13 @@ impl ClientOpcodeMessage {
     pub async fn astd_read<R: async_std::io::ReadExt + Unpin + Send>(mut r: R) -> std::result::Result<Self, crate::errors::ExpectedOpcodeError> {
         let opcode = crate::util::astd_read_u8_le(&mut r).await?;
         match opcode {
-            0x00 => Ok(Self::CMD_AUTH_LOGON_CHALLENGE(CMD_AUTH_LOGON_CHALLENGE_Client::astd_read(r).await?)),
-            0x01 => Ok(Self::CMD_AUTH_LOGON_PROOF(CMD_AUTH_LOGON_PROOF_Client::astd_read(r).await?)),
-            0x02 => Ok(Self::CMD_AUTH_RECONNECT_CHALLENGE(CMD_AUTH_RECONNECT_CHALLENGE_Client::astd_read(r).await?)),
-            0x04 => Ok(Self::CMD_SURVEY_RESULT(CMD_SURVEY_RESULT::astd_read(r).await?)),
-            0x10 => Ok(Self::CMD_REALM_LIST(CMD_REALM_LIST_Client::astd_read(r).await?)),
+            0x00 => Ok(Self::CMD_AUTH_LOGON_CHALLENGE(CMD_AUTH_LOGON_CHALLENGE_Client::astd_read::<R, crate::private::Internal>(r).await?)),
+            0x01 => Ok(Self::CMD_AUTH_LOGON_PROOF(CMD_AUTH_LOGON_PROOF_Client::astd_read::<R, crate::private::Internal>(r).await?)),
+            0x02 => Ok(Self::CMD_AUTH_RECONNECT_CHALLENGE(CMD_AUTH_RECONNECT_CHALLENGE_Client::astd_read::<R, crate::private::Internal>(r).await?)),
+            0x04 => Ok(Self::CMD_SURVEY_RESULT(CMD_SURVEY_RESULT::astd_read::<R, crate::private::Internal>(r).await?)),
+            0x10 => Ok(Self::CMD_REALM_LIST(CMD_REALM_LIST_Client::astd_read::<R, crate::private::Internal>(r).await?)),
             0x32 => Ok(Self::CMD_XFER_ACCEPT),
-            0x33 => Ok(Self::CMD_XFER_RESUME(CMD_XFER_RESUME::astd_read(r).await?)),
+            0x33 => Ok(Self::CMD_XFER_RESUME(CMD_XFER_RESUME::astd_read::<R, crate::private::Internal>(r).await?)),
             0x34 => Ok(Self::CMD_XFER_CANCEL),
             opcode => Err(crate::errors::ExpectedOpcodeError::Opcode(opcode as u32)),
         }